update: moved AnimateThatSearch class to new file in visual effects and commented the search function
